neovim: fix kill buffer mapping
kill buffer mapping for index in 1-9 included the <cr>, so trying to
close buffer > 9 actually killed the ones in 1-9 only (because it was
triggered by the first decimal).

Removed the <cr> at the end of the mapping fixes it.

(I am considering to go back to previous kill buffer "xx", because i
still can't get to get accustomed to the new xb and xx now kills the
whole application)