tmux: add session shortcuts for new and kill
Add new-session binding (to S, uppercase s) to create the session and
renamed it later. This prevents tmux to add a dash-number at the end of
the desired name and also pushes me to give proper name to all sessions.

Add kill-session binding (to K, uppercase k), to kill a session given
its name.