dot: add dot configuration file
The main dot configuration file, which export DOTFILES path and sources
all the aliases in cconf/aliases directory was missing (I probably moved
the code from the zshrc to a dedicated file).

Added it back using dot name (dot for folder and dot.sh for script) to
avoid confusion with aliases/dotfiles which contains only the aliases to
use the dot|dotfile commands.