taskwarrior: update contexts
Currently I am using a major and minor project name.

hsw is the main prject for work related projects, while "me" is the one
related to personal projects.

This is the best way I have found so far to simplify context creation,
that is, I do not have to continuously add filters to context to
differentiate between personal and work releated projects, but it is
also quite time consuming considering that any new project (or even
modify an existing one) requires this prefix.

Looking for a streamline way to achieve the same result.