nvim: use C-l/h to move word by word
I use `w` as prefix to move between splits, to avoid moving my hands to
reach Ctrl (basically w is mapped to C-w), so using w to move to the
next begin of a word is slow, waiting for a longer shortcut to be
activated.

Since I already use also C-l to move in a line by 10 characters always,
I am going to try to use it to move to the next word instead.

C-h will be used in place of b as well, even if b is not remapped to
anything right now.