Add dot.sh script
Currently I use --git-dir flag, and some more configuration, to set the
HOME directory as source folder for this repository. This has the
advantage to simplify deploying of dotfiles and basic management, but
when it comes to branches, and some more advanced git commands this is
trickier. Also I'd like to use helper tools like tig, which is excluded
from the current dot configuration.

This script dot.sh create symlinks for the files under revision, so that
i can use this folder as a normal git repository, with all the default
git command supported.