nvim: fix conceal text in normal mode
I discovered that there is a concealcursor configuration that tells
neovim/vim if show concealed text in each mode (normal, command, insert,
visual).
This configuration might have changed in a recent past, because it
wasn't set manually and the default was to show the text in normal mode
when the cursors was on it. I though it was better not to show it - it's
a bit annoying to see text moving - hover it's much worse if not visible
in normal mode because it can be edited only in insert mode, that is
without any vim power.