nvim: remap w to e
w maps to "move to the beginning of next word", which is useful, but I
prefer use w for other things (moving among windows for example), so I
can only use "e", which is the same but points to the end of next word.
Since the latter isn't as useful, simply remap w to e, so that e points
to the beginning of the next word.