nvim: free "w" from mapping
currently "w" is the start char for some mappings (e.g. move between
splits), so it cannot be used to jump to the "next word" which is the
default behaviour.

To test if it is better to use "w" as default, I replaced it with "s" to
jumpt between splits (mnemonic _s_plits), and to save (mnemonic _s_ave).