nvim: simplify "rep" map
"rep" shortcut is used to feed the commands to replace text.
Currently it sets the flags "gc" which means replace globally and ask
for each replace. Recently I learnt however a couple of new things that
makes this unnecessary sometimes. In particular:

%s//word replace what's inside the * register
%s/word/replace   replace everything without asking and with other
plugins I already see the preview so it's fine if it doesn't ask.