Limit lsp show documentation in insert mode to golang
The markdown lsp is annoing with this configuration, as it constantly
show the links to current page (the "documentation" for the current
page), in case this option is enabled.
However, this option is useful in go when the current line also has an
error. In fact the floating window about the error keeps overriding the
floating window showing the doc if we only use the mapping in normal
mode (which is capital k). If show documentation in insert mode is
enabled, this will allow to see the doc even in the above case.