tmux letsdo: add a 15 minutes reminder
Since on KDE Plasma I don't have the equivalent of OneThing Gnome
extension to show system wide what I'm up to on Letsdo, I re-enabled
this tmux plugin to show a reminder each 15 minutes with the time
currently spent on the same task. It is working fine now, I might
consider adjusting the interval in the future.